% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dimreduction_and_logit.R
\name{dimreduce_and_logit}
\alias{dimreduce_and_logit}
\title{Dimensionality Reduction and Logistic Regression}
\usage{
dimreduce_and_logit(train, isscale = TRUE)
}
\arguments{
\item{train}{A data.frame with features and binary outcome \code{grading}.}

\item{isscale}{Whether to standardize X1 and X2. Default is TRUE.}
}
\value{
A list with two marginal distributions and the fitted glm object.
}
\description{
Dimensionality Reduction and Logistic Regression
}
