% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dimreduction_and_logit_LPD.R
\name{dimreduce_and_logit_LPD}
\alias{dimreduce_and_logit_LPD}
\title{Dimensionality Reduction and Classifier Training (with Latent Space SMOTE)}
\usage{
dimreduce_and_logit_LPD(train, isscale = TRUE, classifier = "logistic")
}
\arguments{
\item{train}{A data.frame with features and binary outcome \code{Y}.}

\item{isscale}{Whether to standardize X1 and X2. Default is TRUE.}

\item{classifier}{The type of classifier ('logistic', 'svm', 'rf', 'xgboost').}
}
\value{
A list with two marginal distributions and the fitted model object.
}
\description{
Dimensionality Reduction and Classifier Training (with Latent Space SMOTE)
}
