% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prediction_P4MC.R
\name{get_prediction_P4MC}
\alias{get_prediction_P4MC}
\title{Get prediction for P4MC using KDE and logistic regression}
\usage{
get_prediction_P4MC(
  i,
  test_data,
  S,
  grid,
  kde.y0.d2,
  kde.y0.d3,
  kde.y1.d2,
  kde.y1.d3,
  logit_fit
)
}
\arguments{
\item{i}{The index of the test sample.}

\item{test_data}{The test data.}

\item{S}{The iterations of MCMC sampling.}

\item{grid}{The number of grid points to extract from the KDE in the first dimension.}

\item{kde.y0.d2}{KDE estimates for class 2, differentiation = 2.}

\item{kde.y0.d3}{KDE estimates for class 3, differentiation = 3.}

\item{kde.y1.d2}{KDE estimates for class 2, differentiation = 2 for class 1.}

\item{kde.y1.d3}{KDE estimates for class 3, differentiation = 3 for class 1.}

\item{logit_fit}{The logistic regression model.}
}
\value{
A numeric value representing the predicted probability.
}
\description{
Get prediction for P4MC using KDE and logistic regression
}
