% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_P4MC.R
\name{run_P4MC}
\alias{run_P4MC}
\title{Run P4MC evaluation with Comprehensive F1 Scores + Optional AUC}
\usage{
run_P4MC(
  datasets,
  k = 10,
  S = 200,
  grid = 25,
  seed = 123,
  csv_path = "P4MC_results_comprehensive.csv",
  calculate_auc = FALSE
)
}
\arguments{
\item{datasets}{A data.frame containing the full dataset.}

\item{k}{The Number of folds. Default is 10.}

\item{S}{The iterations of MCMC sampling. Default is 200.}

\item{grid}{The number of grid points. Default is 25.}

\item{seed}{Random seed. Default is 123.}

\item{csv_path}{The file path to save the result CSV.}

\item{calculate_auc}{Logical. Whether to calculate AUC. Default is FALSE.}
}
\value{
A list with detailed scores and the final summary dataframe.
}
\description{
Run P4MC evaluation with Comprehensive F1 Scores + Optional AUC
}
