% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_P4MC_multi_sim.R
\name{run_P4MC_multi_sim}
\alias{run_P4MC_multi_sim}
\title{Run P4MC Evaluation for Multiple Classifiers}
\usage{
run_P4MC_multi_sim(
  datasets,
  k = 10,
  S = 200,
  grid = 25,
  seed = 123,
  calculate_auc = FALSE,
  classifiers = c("logistic", "svm", "rf", "xgboost")
)
}
\arguments{
\item{datasets}{A data.frame containing the full dataset.}

\item{k}{The number of folds. Default is 10.}

\item{S}{The iterations of MCMC sampling. Default is 200.}

\item{grid}{The number of grid points for KDE. Default is 25.}

\item{seed}{Random seed. Default is 123.}

\item{calculate_auc}{Logical. Whether to calculate AUC. Default is FALSE.}

\item{classifiers}{A vector of classifiers to evaluate.}
}
\value{
A list containing the averaged summary table and raw predictions.
}
\description{
Run P4MC Evaluation for Multiple Classifiers
}
